address review comments

+++ span-processing-config-service-impl/src/main/java/org/hypertrace/span/processing/config/service/SpanProcessingConfigServiceImpl.java
@@ -1,5 +1,7 @@
 package org.hypertrace.span.processing.config.service;
 
+import static org.hypertrace.span.processing.config.service.v1.RuleType.RULE_TYPE_SYSTEM;
+
 import com.google.inject.Inject;
 import com.google.protobuf.util.JsonFormat;
 import com.typesafe.config.Config;
@@ -53,15 +55,14 @@ import org.hypertrace.span.processing.config.service.validation.SpanProcessingCo
 class SpanProcessingConfigServiceImpl
     extends SpanProcessingConfigServiceGrpc.SpanProcessingConfigServiceImplBase {
 
-  private static final String SPAN_PROCESSING_CONFIG_SERVICE = "span.processing.config.service";
-  private static final String SYSTEM_LEVEL_EXCLUDE_SPAN_RULES = "system.level.exclude.span.rules";
+  private static final String SYSTEM_EXCLUDE_SPAN_RULES =
+      "span.processing.config.service.system.exclude.span.rules";
 
   private final SpanProcessingConfigRequestValidator validator;
   private final ExcludeSpanRulesConfigStore excludeSpanRulesConfigStore;
   private final TimestampConverter timestampConverter;
   private final ApiNamingRulesManager apiNamingRulesManager;
-  private List<ExcludeSpanRule> systemLevelExcludeSpanRules;
-  private Map<String, ExcludeSpanRuleDetails> systemLevelExcludeSpanRulesToIdMap;
+  private Map<String, ExcludeSpanRule> systemExcludeSpanRuleIdToRuleMap;
 
   @Inject
   SpanProcessingConfigServiceImpl(
@@ -75,25 +76,19 @@ class SpanProcessingConfigServiceImpl
     this.timestampConverter = timestampConverter;
     this.apiNamingRulesManager = apiNamingRulesManager;
 
-    buildSystemLevelExcludeSpanRuleConfigs(config);
+    buildSystemExcludeSpanRuleConfigs(config);
   }
 
-  private void buildSystemLevelExcludeSpanRuleConfigs(Config config) {
-    List<? extends com.typesafe.config.ConfigObject> systemLevelExcludeSpanRuleObjectList = null;
-    if (config.hasPath(SPAN_PROCESSING_CONFIG_SERVICE)) {
-      Config spanProcessingConfig = config.getConfig(SPAN_PROCESSING_CONFIG_SERVICE);
-      if (spanProcessingConfig.hasPath(SYSTEM_LEVEL_EXCLUDE_SPAN_RULES)) {
-        systemLevelExcludeSpanRuleObjectList =
-            spanProcessingConfig.getObjectList(SYSTEM_LEVEL_EXCLUDE_SPAN_RULES);
-      }
+  private void buildSystemExcludeSpanRuleConfigs(Config config) {
+    List<? extends com.typesafe.config.ConfigObject> systemExcludeSpanRuleObjectList = null;
+    if (config.hasPath(SYSTEM_EXCLUDE_SPAN_RULES)) {
+      systemExcludeSpanRuleObjectList = config.getObjectList(SYSTEM_EXCLUDE_SPAN_RULES);
     }
-    if (systemLevelExcludeSpanRuleObjectList != null) {
-      systemLevelExcludeSpanRules =
-          buildSystemLevelExcludeSpanRules(systemLevelExcludeSpanRuleObjectList);
-      systemLevelExcludeSpanRulesToIdMap = buildSystemLevelExcludeSpanRulesToIdMap();
+    if (systemExcludeSpanRuleObjectList != null) {
+      systemExcludeSpanRuleIdToRuleMap =
+          buildSystemExcludeSpanRulesToIdMap(systemExcludeSpanRuleObjectList);
     } else {
-      systemLevelExcludeSpanRules = Collections.emptyList();
-      systemLevelExcludeSpanRulesToIdMap = Collections.emptyMap();
+      systemExcludeSpanRuleIdToRuleMap = Collections.emptyMap();
     }
   }
 
@@ -107,32 +102,32 @@ class SpanProcessingConfigServiceImpl
       List<ExcludeSpanRuleDetails> excludeSpanRules =
           excludeSpanRulesConfigStore.getAllData(requestContext);
 
-      Map<String, ExcludeSpanRuleDetails> tenantLevelExcludeSpanRulesToIdMap =
-          buildTenantLevelExcludeSpanRulesToIdMap(excludeSpanRules);
+      Map<String, ExcludeSpanRuleDetails> userExcludeSpanRuleIdToRuleMap =
+          buildUserExcludeSpanRulesToIdMap(excludeSpanRules);
 
-      // all the tenant level configs except for the overridden system level ones
-      List<ExcludeSpanRuleDetails> filteredTenantLevelExcludeSpanRules =
+      // all the user configs except for the overridden system ones
+      List<ExcludeSpanRuleDetails> filteredUserExcludeSpanRules =
           excludeSpanRules.stream()
               .filter(
                   excludeSpanRule ->
-                      !systemLevelExcludeSpanRulesToIdMap.containsKey(
+                      !systemExcludeSpanRuleIdToRuleMap.containsKey(
                           excludeSpanRule.getRule().getId()))
               .collect(Collectors.toUnmodifiableList());
 
-      // all the system level configs, replaced by the overridden configs in case of overrides
-      List<ExcludeSpanRuleDetails> filteredSystemLevelExcludeSpanRules =
-          systemLevelExcludeSpanRules.stream()
+      // all the system configs, replaced by the overridden configs in case of overrides
+      List<ExcludeSpanRuleDetails> filteredSystemExcludeSpanRules =
+          systemExcludeSpanRuleIdToRuleMap.values().stream()
               .map(
                   excludeSpanRule ->
-                      tenantLevelExcludeSpanRulesToIdMap.containsKey(excludeSpanRule.getId())
-                          ? tenantLevelExcludeSpanRulesToIdMap.get(excludeSpanRule.getId())
+                      userExcludeSpanRuleIdToRuleMap.containsKey(excludeSpanRule.getId())
+                          ? userExcludeSpanRuleIdToRuleMap.get(excludeSpanRule.getId())
                           : ExcludeSpanRuleDetails.newBuilder().setRule(excludeSpanRule).build())
               .collect(Collectors.toUnmodifiableList());
 
       responseObserver.onNext(
           GetAllExcludeSpanRulesResponse.newBuilder()
-              .addAllRuleDetails(filteredTenantLevelExcludeSpanRules)
-              .addAllRuleDetails(filteredSystemLevelExcludeSpanRules)
+              .addAllRuleDetails(filteredUserExcludeSpanRules)
+              .addAllRuleDetails(filteredSystemExcludeSpanRules)
               .build());
       responseObserver.onCompleted();
     } catch (Exception e) {
@@ -179,13 +174,17 @@ class SpanProcessingConfigServiceImpl
 
       UpdateExcludeSpanRule updateExcludeSpanRule = request.getRule();
 
-      // check if the rule already exists. If not, check if it is a system level config. If yes, use
+      // check if the rule already exists. If not, check if it is a system config. If yes, use
       // that. Else, error out.
       ExcludeSpanRule existingRule =
           this.excludeSpanRulesConfigStore
               .getData(requestContext, updateExcludeSpanRule.getId())
-              .or(() -> getSystemLevelExcludeSpanRule(updateExcludeSpanRule.getId()))
+              .or(() -> getSystemExcludeSpanRule(updateExcludeSpanRule.getId()))
               .orElseThrow(Status.NOT_FOUND::asException);
+
+      // if the rule being updated is a system exclude rule, only disabled field update should be
+      // allowed
+      validateUpdateExcludeSpanRule(existingRule, updateExcludeSpanRule);
       ExcludeSpanRule updatedRule = buildUpdatedRule(existingRule, updateExcludeSpanRule);
 
       responseObserver.onNext(
@@ -209,6 +208,11 @@ class SpanProcessingConfigServiceImpl
       RequestContext requestContext = RequestContext.CURRENT.get();
       this.validator.validateOrThrow(requestContext, request);
 
+      // do not allow deleting system exclude rules
+      if (systemExcludeSpanRuleIdToRuleMap.containsKey(request.getId())) {
+        throw Status.INVALID_ARGUMENT.asRuntimeException();
+      }
+
       // TODO: need to handle priorities
       this.excludeSpanRulesConfigStore
           .deleteObject(requestContext, request.getId())
@@ -222,13 +226,6 @@ class SpanProcessingConfigServiceImpl
     }
   }
 
-  private List<ExcludeSpanRule> buildSystemLevelExcludeSpanRules(
-      List<? extends com.typesafe.config.ConfigObject> configObjectList) {
-    return configObjectList.stream()
-        .map(this::buildExcludeSpanRuleFromConfig)
-        .collect(Collectors.toUnmodifiableList());
-  }
-
   @SneakyThrows
   private ExcludeSpanRule buildExcludeSpanRuleFromConfig(
       com.typesafe.config.ConfigObject configObject) {
@@ -238,29 +235,38 @@ class SpanProcessingConfigServiceImpl
     return builder.build();
   }
 
-  private Optional<ExcludeSpanRule> getSystemLevelExcludeSpanRule(String id) {
-    if (systemLevelExcludeSpanRulesToIdMap.containsKey(id)) {
-      return Optional.of(systemLevelExcludeSpanRulesToIdMap.get(id).getRule());
+  private Optional<ExcludeSpanRule> getSystemExcludeSpanRule(String id) {
+    if (systemExcludeSpanRuleIdToRuleMap.containsKey(id)) {
+      return Optional.of(systemExcludeSpanRuleIdToRuleMap.get(id));
     }
     return Optional.empty();
   }
 
-  private Map<String, ExcludeSpanRuleDetails> buildSystemLevelExcludeSpanRulesToIdMap() {
-    return systemLevelExcludeSpanRules.stream()
-        .collect(
-            Collectors.toUnmodifiableMap(
-                ExcludeSpanRule::getId,
-                (ExcludeSpanRule excludeSpanRule) ->
-                    ExcludeSpanRuleDetails.newBuilder().setRule(excludeSpanRule).build()));
+  private Map<String, ExcludeSpanRule> buildSystemExcludeSpanRulesToIdMap(
+      List<? extends com.typesafe.config.ConfigObject> configObjectList) {
+    return configObjectList.stream()
+        .map(this::buildExcludeSpanRuleFromConfig)
+        .collect(Collectors.toUnmodifiableMap(ExcludeSpanRule::getId, Function.identity()));
   }
 
-  private Map<String, ExcludeSpanRuleDetails> buildTenantLevelExcludeSpanRulesToIdMap(
+  private Map<String, ExcludeSpanRuleDetails> buildUserExcludeSpanRulesToIdMap(
       List<ExcludeSpanRuleDetails> excludeSpanRules) {
     return excludeSpanRules.stream()
         .collect(
             Collectors.toUnmodifiableMap(
                 (ExcludeSpanRuleDetails ruleDetails) -> ruleDetails.getRule().getId(),
                 Function.identity()));
+  }
+
+  private void validateUpdateExcludeSpanRule(
+      ExcludeSpanRule existingRule, UpdateExcludeSpanRule updateExcludeSpanRule) {
+    ExcludeSpanRuleInfo ruleInfo = existingRule.getRuleInfo();
+    if (RULE_TYPE_SYSTEM.equals(ruleInfo.getType())) {
+      if (!updateExcludeSpanRule.getName().equals(ruleInfo.getName())
+          || !updateExcludeSpanRule.getFilter().equals(ruleInfo.getFilter())) {
+        throw Status.INVALID_ARGUMENT.asRuntimeException();
+      }
+    }
   }
 
   private ExcludeSpanRule buildUpdatedRule(
